requirements.txt
streamlit
requests
beautifulsoup4
ollama
httpx
pandas
PyPDF2
Pillow